+++ changelog.md
@@ -6,6 +6,7 @@
 === Functionality ===
 
 * HPOlib/Plotting/results.py: New result print script which can be used while the experiment is still running
+* HPOlib/Plotting/doFanovaPlots.py;HPOlib/scripts/HPOlib-pyFanova: New plotting script that calls pyfanova and saves plots
 
 === Other ===